+++ src/api/java/at/minecraftschurli/arsmagicalegacy/api/etherium/EtheriumHandlerBlock.java
@@ -4,11 +4,10 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.phys.AABB;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * Represents a block that has an {@link EtheriumHandler} {@link BlockEntity} attached. Methods in this class mainly serve selection or visual purposes.
+ * Represents a block that has an {@link EtheriumHandler} {@link BlockEntity} attached. Methods in this class mainly serve selection purposes.
  */
 public interface EtheriumHandlerBlock {
     /**
@@ -22,16 +21,4 @@ public interface EtheriumHandlerBlock {
      */
     @Nullable
     BlockEntity getBlockEntity(Level level, BlockPos pos, BlockState state);
-
-    /**
-     * Returns the block outline {@link AABB} to render when the Magitech Goggles are equipped.
-     * If this is a block larger than 1x1x1, only the part that actually controls the logic should return an {@link AABB}, all other parts should return null.
-     *
-     * @param level The {@link Level} to use.
-     * @param pos   The {@link BlockPos} to use.
-     * @param state The {@link BlockState} to use.
-     * @return The block outline {@link AABB}.
-     */
-    @Nullable
-    AABB getOutline(Level level, BlockPos pos, BlockState state);
 }