+++ src/main/java/at/minecraftschurli/arsmagicalegacy/block/CelestialPrismBlock.java
@@ -16,7 +16,6 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.EnumProperty;
 import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.Nullable;
@@ -95,12 +94,6 @@ public class CelestialPrismBlock extends EtheriumGeneratorBlock {
             case UPPER -> pos.below(1);
         };
         return level.getBlockState(pos).is(this) && level.getBlockState(pos).getValue(PART) == Part.LOWER && level.getBlockEntity(pos) instanceof CelestialPrismBlockEntity blockEntity ? blockEntity : null;
-    }
-
-    @Override
-    @Nullable
-    public AABB getOutline(Level level, BlockPos pos, BlockState state) {
-        return state.getValue(PART) == Part.LOWER ? new AABB(pos.getX(), pos.getY(), pos.getZ(), pos.getX() + 1, pos.getY() + 2, pos.getZ() + 1) : null;
     }
 
     public enum Part implements StringRepresentableEnum {