+++ src/main/java/at/minecraftschurli/arsmagicalegacy/client/AMRenderTypes.java
@@ -1,5 +1,6 @@
 package at.minecraftschurli.arsmagicalegacy.client;
 
+import at.minecraftschurli.arsmagicalegacy.api.ArsMagicaApi;
 import at.minecraftschurli.arsmagicalegacy.client.atlas.SpellIconAtlasHolder;
 import at.minecraftschurli.arsmagicalegacy.client.gui.spellcustomization.color.ColorWheelShader;
 import com.mojang.blaze3d.shaders.Uniform;
@@ -13,7 +14,7 @@ import net.minecraft.client.renderer.ShaderInstance;
 public final class AMRenderTypes {
     private static final RenderStateShard.ShaderStateShard COLOR_WHEEL_SHADER = new RenderStateShard.ShaderStateShard(ColorWheelShader::getInstance);
     public static final RenderType COLOR_WHEEL = RenderType.create(
-        "color_wheel",
+        ArsMagicaApi.modLoc("color_wheel").toString().replace(":", "_"),
         DefaultVertexFormat.POSITION_COLOR,
         VertexFormat.Mode.QUADS,
         256,
@@ -28,10 +29,29 @@ public final class AMRenderTypes {
                 setUniform("radius", ColorWheelShader.getRadius());
                 setUniform("brightness", ColorWheelShader.getBrightness());
             }, () -> {}))
-            .createCompositeState(false)
-    );
+            .createCompositeState(false));
     public static final RenderType SPELL_ICON = RenderType.itemEntityTranslucentCull(SpellIconAtlasHolder.ATLAS);
     public static final RenderType SPELL_ICON_FABULOUS = RenderType.entityTranslucentCull(SpellIconAtlasHolder.ATLAS);
+    public static final RenderType LINES_WITH_WIDTH = RenderType.create(
+        ArsMagicaApi.modLoc("lines_with_width").toString().replace(":", "_"),
+        DefaultVertexFormat.POSITION_COLOR,
+        VertexFormat.Mode.TRIANGLES,
+        8192,
+        false,
+        false,
+        RenderType.CompositeState.builder()
+            .setTextureState(RenderStateShard.NO_TEXTURE)
+            .setShaderState(RenderStateShard.POSITION_COLOR_SHADER)
+            .setTransparencyState(RenderStateShard.TRANSLUCENT_TRANSPARENCY)
+            .setDepthTestState(RenderStateShard.LEQUAL_DEPTH_TEST)
+            .setCullState(RenderStateShard.CULL)
+            .setLightmapState(RenderStateShard.NO_LIGHTMAP)
+            .setOverlayState(RenderStateShard.NO_OVERLAY)
+            .setLayeringState(RenderStateShard.NO_LAYERING)
+            .setOutputState(RenderStateShard.MAIN_TARGET)
+            .setTexturingState(RenderStateShard.DEFAULT_TEXTURING)
+            .setWriteMaskState(RenderStateShard.COLOR_DEPTH_WRITE)
+            .createCompositeState(false));
 
     private static void setUniform(String name, float value) {
         ShaderInstance shader = RenderSystem.getShader();