+++ src/main/java/at/minecraftschurli/arsmagicalegacy/item/MagitechGogglesItem.java
@@ -3,6 +3,7 @@ package at.minecraftschurli.arsmagicalegacy.item;
 import at.minecraftschurli.arsmagicalegacy.api.ArsMagicaApi;
 import at.minecraftschurli.arsmagicalegacy.init.AMItems;
 import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -24,5 +25,9 @@ public class MagitechGogglesItem extends ArmorItem {
 
     public MagitechGogglesItem(Properties properties) {
         super(AMItems.MAGITECH_GOGGLES_MATERIAL, Type.HELMET, properties.stacksTo(1));
+    }
+
+    public static boolean shouldRender(Player player) {
+        return player.getInventory().getArmor(3).is(AMItems.MAGITECH_GOGGLES);
     }
 }